WIP start adding support for processes and their sequential statements. Also lay groundwork for relevant types.
