+++ crates/vhdl/src/process/statement/case.rs
@@ -1,0 +1,2 @@
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct Case {}
