+++ crates/vhdl/src/process/statement/condition.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub enum Condition {
+    Constant(bool),
+}
