+++ crates/vhdl/src/process/statement/loop_statement.rs
@@ -1,0 +1,7 @@
+use super::condition::Condition;
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub enum LoopStatement {
+    While(Condition),
+    For,
+}
