+++ crates/vhdl/src/process/statement/wait.rs
@@ -1,0 +1,7 @@
+use super::condition::Condition;
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub enum Wait {
+    Wait,
+    WaitUntil(Condition),
+}
